Struct holochain_test_bin::p2p_node::test_node::TestNode

pub struct TestNode {
    pub config: P2pConfig,
    pub agent_id: Address,
    pub current_dna: Option<Address>,
    pub logger: TweetProxy,
    pub p2p_binding: String,
    // some fields omitted
}

Conductor Mock of one agent with multiple DNAs

Fields
config: P2pConfigagent_id: Addresscurrent_dna: Option<Address>logger: TweetProxyp2p_binding: String
Methods

impl TestNode

Query logs

pub fn count_recv_json_messages(&self) -> usize

Return number of JsonProtocol message this node has received

pub fn find_recv_json_msg(
&self,
ith: usize,
predicate: Box<dyn Fn(&JsonProtocol) -> bool>
) -> Option<JsonProtocol>

Return the ith JSON message that this node has received and fullfills predicate

impl TestNode

pub fn track_current_dna(&mut self) -> NetResult<()>

pub fn track_dna(
&mut self,
dna_address: &Address,
can_set_current: bool
) -> NetResult<()>

pub fn untrack_current_dna(&mut self) -> NetResult<()>

pub fn untrack_dna(&mut self, dna_address: &Address) -> NetResult<()>

pub fn is_tracking(&self, dna_address: &Address) -> bool

pub fn set_current_dna(&mut self, dna_address: &Address)

impl TestNode

pub fn author_entry(
&mut self,
entry_address: &Address,
aspect_content_list: Vec<Vec<u8>>,
can_broadcast: bool
) -> NetResult<()>

pub fn hold_entry(
&mut self,
entry_address: &Address,
aspect_content_list: Vec<Vec<u8>>
) -> NetResult<()>

impl TestNode

Query & Fetch

pub fn request_entry(&mut self, entry_address: Address) -> QueryEntryData

Node asks for some entry on the network.

pub fn reply_to_HandleQueryEntry(
&mut self,
query: &QueryEntryData
) -> Result<QueryEntryResultData, GenericResultData>

pub fn reply_to_HandleFetchEntry(
&mut self,
fetch: &FetchEntryData
) -> Result<FetchEntryResultData, GenericResultData>

impl TestNode

pub fn send_direct_message(
&mut self,
to_agent_id: &Address,
content: Vec<u8>
) -> String

Node sends Message on the network.

pub fn send_reponse_json(&mut self, msg: MessageData, response_content: Vec<u8>)

Node sends Message on the network.

pub fn send_reponse_lib3h(
&mut self,
msg: DirectMessageData,
response_content: Value
)

Node sends Message on the network.

impl TestNode

Reply LISTS

pub fn reply_to_HandleGetAuthoringEntryList(
&mut self,
request: &GetListData
) -> NetResult<()>

Reply to a HandleGetAuthoringEntryList request

pub fn reply_to_first_HandleGetAuthoringEntryList(&mut self)

Look for the first HandleGetAuthoringEntryList request received from network module and reply

pub fn reply_to_HandleGetHoldingEntryList(
&mut self,
request: &GetListData
) -> NetResult<()>

Reply to a HandleGetHoldingEntryList request

pub fn reply_to_first_HandleGetHoldingEntryList(&mut self)

Look for the first HandleGetHoldingEntryList request received from network module and reply

impl TestNode

pub fn new_with_config(
agent_id_arg: Address,
config: &P2pConfig,
_maybe_temp_dir: Option<TempDir>
) -> Self

Private constructor

pub fn is_network_ready(&self) -> bool

pub fn new_with_unique_memory_network(agent_id: Address) -> Self

Constructor for an in-memory P2P Network

pub fn new_with_uri_ipc_network(agent_id: Address, ipc_binding: &str) -> Self

Constructor for an IPC node that uses an existing n3h process and a temp folder

pub fn new_with_lib3h(
agent_id: Address,
maybe_config_filepath: Option<&str>,
maybe_end_user_config_filepath: Option<String>,
bootstrap_nodes: Vec<String>,
maybe_dir_path: Option<String>
) -> Self

Constructor for an IPC node that uses an existing n3h process and a temp folder

pub fn new_with_spawn_ipc_network(
agent_id: Address,
maybe_config_filepath: Option<&str>,
maybe_end_user_config_filepath: Option<String>,
bootstrap_nodes: Vec<String>,
maybe_dir_path: Option<String>
) -> Self

Constructor for an IPC node that spawns and uses a n3h process and a temp folder

pub fn try_recv(&mut self) -> NetResult<Protocol>

pub fn try_recv_json(&mut self) -> NetResult<JsonProtocol>

See if there is a message to receive, and log it return a JsonProtocol if the received message is of that type

pub fn try_recv_lib3h(&mut self) -> NetResult<Lib3hServerProtocol>

See if there is a message to receive, and log it return a JsonProtocol if the received message is of that type

pub fn listen(&mut self, timeout_ms: usize) -> usize

recv messages until timeout is reached returns the number of messages it received during listening period timeout is reset after a message is received

pub fn wait_HandleFetchEntry_and_reply(&mut self) -> bool

wait to receive a HandleFetchEntry request and automatically reply return true if a HandleFetchEntry has been received

pub fn wait_HandleQueryEntry_and_reply(&mut self) -> bool

wait to receive a HandleFetchEntry request and automatically reply return true if a HandleFetchEntry has been received

pub fn wait_json_with_timeout(
&mut self,
predicate: Box<dyn Fn(&JsonProtocol) -> bool>,
timeout_ms: usize
) -> Option<JsonProtocol>

Wait for receiving a message corresponding to predicate until timeout is reached

pub fn wait_json(
&mut self,
predicate: Box<dyn Fn(&JsonProtocol) -> bool>
) -> Option<JsonProtocol>

Wait for receiving a message corresponding to predicate hard coded timeout

pub fn wait_lib3h(
&mut self,
predicate: Box<dyn Fn(&Lib3hServerProtocol) -> bool>
) -> Option<Lib3hServerProtocol>

Wait for receiving a message corresponding to predicate hard coded timeout

pub fn wait_lib3h_with_timeout(
&mut self,
predicate: Box<dyn Fn(&Lib3hServerProtocol) -> bool>,
timeout_ms: usize
) -> Option<Lib3hServerProtocol>

Wait for receiving a message corresponding to predicate until timeout is reached

pub fn stop(self)

pub fn endpoint(&self) -> String

Getter of the endpoint of its connection