Struct holochain_test_bin::p2p_node::test_node::TestNode

pub struct TestNode {
    pub config: P2pConfig,
    pub agent_id: String,
    pub current_dna: Option<Address>,
    pub logger: TweetProxy,
    pub p2p_binding: String,
    // some fields omitted
}

Core Mock

Fields
config: P2pConfigagent_id: Stringcurrent_dna: Option<Address>logger: TweetProxyp2p_binding: String
Methods

impl TestNode

Query logs

pub fn count_recv_json_messages(&self) -> usize

Return number of JsonProtocol message this node has received

pub fn find_recv_msg(
&self,
ith: usize,
predicate: Box<dyn Fn(&JsonProtocol) -> bool>
) -> Option<JsonProtocol>

Return the ith JSON message that this node has received and fullfills predicate

impl TestNode

pub fn track_current_dna(&mut self) -> NetResult<()>

pub fn track_dna(
&mut self,
dna_address: &Address,
can_set_current: bool
) -> NetResult<()>

pub fn untrack_current_dna(&mut self) -> NetResult<()>

pub fn untrack_dna(&mut self, dna_address: &Address) -> NetResult<()>

pub fn is_tracking(&self, dna_address: &Address) -> bool

pub fn set_current_dna(&mut self, dna_address: &Address)

impl TestNode

publish, hold

pub fn author_entry(
&mut self,
entry_address: &Address,
entry_content: &Value,
can_publish: bool
) -> NetResult<()>

pub fn author_meta(
&mut self,
entry_address: &Address,
attribute: &str,
link_entry_address: &Value,
can_publish: bool
) -> NetResult<MetaKey>

pub fn hold_entry(&mut self, entry_address: &Address, entry_content: &Value)

pub fn hold_meta(
&mut self,
entry_address: &Address,
attribute: &str,
link_entry_address: &Value
)

impl TestNode

fetch & sendMessage

pub fn request_entry(&mut self, entry_address: Address) -> FetchEntryData

Node asks for some entry on the network.

pub fn request_meta(
&mut self,
entry_address: Address,
attribute: String
) -> FetchMetaData

Node asks for some meta on the network.

pub fn send_message(
&mut self,
to_agent_id: String,
content: Value
) -> MessageData

Node sends Message on the network.

pub fn send_reponse(
&mut self,
msg: MessageData,
response_content: Value
) -> MessageData

Node sends Message on the network.

impl TestNode

pub fn reply_to_HandleFetchEntry(
&mut self,
request: &FetchEntryData
) -> NetResult<()>

Send a reponse to a FetchDhtData request

pub fn reply_to_HandleFetchMeta(
&mut self,
request: &FetchMetaData
) -> NetResult<()>

Send a reponse to a FetchDhtMetaData request

pub fn reply_to_HandleGetPublishingEntryList(
&mut self,
request: &GetListData
) -> NetResult<()>

Reply to a HandleGetPublishingEntryList request

pub fn reply_to_first_HandleGetPublishingEntryList(&mut self)

Look for the first HandleGetPublishingEntryList request received from network module and reply

pub fn reply_to_HandleGetPublishingMetaList(
&mut self,
request: &GetListData
) -> NetResult<()>

Reply to a HandleGetPublishingMetaList request

pub fn reply_to_first_HandleGetPublishingMetaList(&mut self)

Look for the first HandleGetPublishingMetaList request received from network module and reply

pub fn reply_to_HandleGetHoldingEntryList(
&mut self,
request: &GetListData
) -> NetResult<()>

Reply to a HandleGetHoldingEntryList request

pub fn reply_to_first_HandleGetHoldingEntryList(&mut self)

Look for the first HandleGetHoldingEntryList request received from network module and reply

pub fn reply_to_HandleGetHoldingMetaList(
&mut self,
request: &GetListData
) -> NetResult<()>

Reply to a HandleGetHoldingMetaList request

pub fn reply_to_first_HandleGetHoldingMetaList(&mut self)

Look for the first HandleGetHoldingMetaList request received from network module and reply

impl TestNode

pub fn new_with_config(
agent_id_arg: String,
config: &P2pConfig,
_maybe_temp_dir: Option<TempDir>
) -> Self

Private constructor

pub fn is_network_ready(&self) -> bool

pub fn new_with_unique_memory_network(agent_id: String) -> Self

Constructor for an in-memory P2P Network

pub fn new_with_uri_ipc_network(agent_id: String, ipc_binding: &str) -> Self

Constructor for an IPC node that uses an existing n3h process and a temp folder

pub fn new_with_spawn_ipc_network(
agent_id: String,
maybe_config_filepath: Option<&str>,
maybe_end_user_config_filepath: Option<String>,
bootstrap_nodes: Vec<String>,
maybe_dir_path: Option<String>
) -> Self

Constructor for an IPC node that spawns and uses a n3h process and a temp folder

pub fn try_recv(&mut self) -> NetResult<JsonProtocol>

See if there is a message to receive, and log it return a JsonProtocol if the received message is of that type

pub fn listen(&mut self, timeout_ms: usize) -> usize

recv messages until timeout is reached returns the number of messages it received during listening period timeout is reset after a message is received

pub fn wait_HandleFetchEntry_and_reply(&mut self) -> bool

wait to receive a HandleFetchEntry request and automatically reply return true if a HandleFetchEntry has been received

pub fn wait_HandleFetchMeta_and_reply(&mut self) -> bool

wait to receive a HandleFetchMeta request and automatically reply return true if a HandleFetchMeta has been received

pub fn wait_with_timeout(
&mut self,
predicate: Box<dyn Fn(&JsonProtocol) -> bool>,
timeout_ms: usize
) -> Option<JsonProtocol>

Wait for receiving a message corresponding to predicate until timeout is reached

pub fn wait(
&mut self,
predicate: Box<dyn Fn(&JsonProtocol) -> bool>
) -> Option<JsonProtocol>

Wait for receiving a message corresponding to predicate hard coded timeout

pub fn stop(self)

pub fn endpoint(&self) -> String

Getter of the endpoint of its connection